backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0
aiofiles>=23.0.0
websockets>=12.0
pydantic>=2.0.0
twelvelabs>=0.2.0
yt-dlp>=2024.3.10
requests